requirements.txt
alembic==1.13.0
autopep8==2.0.4
blinker==1.7.0
click==8.1.7
colorama==0.4.6
cssbeautifier==1.15.1
djlint==1.34.1
EditorConfig==0.12.4
Flask==3.0.0
Flask-Admin==1.6.1
Flask-CKEditor==0.5.1
Flask-Email==1.4.4
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
flask-paginate==2024.3.28
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.2
gunicorn==22.0.0
html-tag-names==0.1.2
html-void-elements==0.1.0
importlib-metadata==7.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsbeautifier==1.15.1
json5==0.9.24
Mako==1.3.0
MarkupSafe==2.1.3
packaging==24.0
pathspec==0.12.1
pillow==10.3.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycodestyle==2.11.1
python-decouple==3.8
PyYAML==6.0.1
regex==2023.12.25
six==1.16.0
SQLAlchemy==2.0.23
tomli==2.0.1
tqdm==4.66.2
typing_extensions==4.9.0
Werkzeug==3.0.1
WTForms==3.1.2
zipp==3.17.0
